Fix netsnmp_init_mib configure test
I added this test to see if the system has the netsnmp_init_mib()
function.  If it does not, we fall back to the older init_mib().
Unfortunately, the test always passes, because it only tried to compile,
and did not link.

Now, it links, so it works correctly.  It now also prints status output.

This closes #44.